add numeric sensor for mysql

Related work items: #6772

+++ dqoai/src/integration-test/java/ai/dqo/mysql/column/numeric/MysqlColumnNumericInvalidLatitudeCountSensorParametersSpecIntegrationTest.java
@@ -64,7 +64,7 @@ public class MysqlColumnNumericInvalidLatitudeCountSensorParametersSpecIntegrati
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(13f, resultTable.column(0).get(0));
+        Assertions.assertEquals(13, (double) resultTable.column(0).get(0));
     }
 
     @Test
@@ -77,7 +77,7 @@ public class MysqlColumnNumericInvalidLatitudeCountSensorParametersSpecIntegrati
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(13f, resultTable.column(0).get(0));
+        Assertions.assertEquals(13, (double) resultTable.column(0).get(0));
     }
 
     @Test
@@ -90,7 +90,7 @@ public class MysqlColumnNumericInvalidLatitudeCountSensorParametersSpecIntegrati
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(1, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(13f, resultTable.column(0).get(0));
+        Assertions.assertEquals(13, (double) resultTable.column(0).get(0));
     }
 
     @Test
@@ -103,7 +103,7 @@ public class MysqlColumnNumericInvalidLatitudeCountSensorParametersSpecIntegrati
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(4, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(5f, resultTable.column(0).get(0));
+        Assertions.assertEquals(5, (double) resultTable.column(0).get(0));
     }
 
     @Test
@@ -116,6 +116,6 @@ public class MysqlColumnNumericInvalidLatitudeCountSensorParametersSpecIntegrati
         Table resultTable = sensorResult.getResultTable();
         Assertions.assertEquals(4, resultTable.rowCount());
         Assertions.assertEquals("actual_value", resultTable.column(0).name());
-        Assertions.assertEquals(5f, resultTable.column(0).get(0));
+        Assertions.assertEquals(5, (double) resultTable.column(0).get(0));
     }
 }
